[mcache] fixing tests and feedback
