+++ mcache/tox.ini
@@ -2,13 +2,22 @@
 minversion = 2.0
 basepython = py27
 envlist =
-    mcache
+    unit
+    integration
     flake8
 
 [testenv]
 platform = linux|darwin|win32
 
-[testenv:mcache]
+[testenv:unit]
+deps =
+    ../datadog_checks_base
+    -rrequirements-dev.txt
+commands =
+    pip install --require-hashes -r requirements.txt
+    pytest -m"not integration" -v
+
+[testenv:integration]
 passenv =
     DOCKER*
     COMPOSE*
@@ -17,7 +26,7 @@ deps =
     -rrequirements-dev.txt
 commands =
     pip install --require-hashes -r requirements.txt
-    pytest -v
+    pytest -m"integration" -v
 
 [testenv:flake8]
 skip_install = true
